+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Changelog
 
+### v0.1.6 - 2024/04/26
+- Added temple mirroring
+
 ### v0.1.5 - 2024/04/10
 
-- Updated the specification, now requiring nose pads nodes, otherwise loosening requirements
+- Updated the specification loosening requirements
 - Improved temple group creation so that it allows rotation